Modify app payloads

+++ app/products/views.py
@@ -6,11 +6,9 @@ from app.auth.views import check_admin
 import re
 
 
-
-# create product blueprint
-# blueprint will handle all product routes for the app
+#create product blueprint
+#blueprint will handle all product routes for the app
 product = Blueprint('product',__name__)
-
 
 
 @product.route('/products', methods=['POST'])
@@ -29,7 +27,7 @@ def create_product(current_user):
         product_quantity = data['product_quantity']
         product_price = data['product_price']
         product_description = data['product_description']
-        if check_admin(current_user) == True:
+        if check_admin(current_user) is True:
             return jsonify({'error':'Access Denied. Please login as admin'}),401
         #check if content type is application/json
         if not request.content_type == 'application/json': 
@@ -38,7 +36,7 @@ def create_product(current_user):
             return jsonify({'error':'One of the required fields is empty'}),400
         if not isinstance(product_price,int) or not isinstance(product_quantity,int) or product_price < 1 or product_quantity < 1:
             return jsonify({'error':'price or quantity must be a number and must be greater than 1'}),400
-        if re.search('[\s]',product_name) != None:
+        if re.search(r'[\s]',product_name) != None:
             return jsonify({'error':'poductname cannot contain spaces'}),400
         products = db.select_products()
         for product in products:
@@ -60,9 +58,12 @@ def get_products(current_user):
     """
     products = db.select_products()
     if len(products) == 0:
-        return jsonify({'error':'There no products at the moment'}),404
-    return jsonify({'Products-Available':products}),200
-
+        return jsonify({'message':'There no products at the moment'}),404
+    keys = ['Product_id','Product_name','Product_quantity','Product_price','Product_description','Date_added']
+    products_list = []
+    for product in products:
+        products_list.append(dict(zip(keys,product)))
+    return jsonify({'Products':products_list}),200
 
 @product.route('/products/<productid>', methods=['GET'])
 @protected_route
@@ -76,9 +77,16 @@ def get_product(current_user,productid):
         return jsonify({'error':'There no products at the moment'}),404
     product_fetched = db.select_a_product(productid)
     if product_fetched == None:
-        return jsonify({'error':'Product was not found'}),404
-    return jsonify({'result':product_fetched}),200
-
+        return jsonify({'message':'Product was not found'}),404
+    product = {
+              'Product_id':product_fetched[0],
+              'Product_name':product_fetched[1],
+              'Product_quantity':product_fetched[2],
+              'Product_price': product_fetched[3],
+              'Product_description':product_fetched[4],
+              'Date_added':product_fetched[5]
+              }
+    return jsonify({'Product':product}),200
 
 @product.route('/products/<productid>', methods=['PUT'])
 @protected_route
@@ -94,32 +102,43 @@ def update_product(current_user,productid):
         prodt_quantity = data['productquantity']
         prodt_price = data['productprice']
         prodt_desc = data['productdescription']
+        if check_admin(current_user) is True:
+            return jsonify({'error':'Access Denied. Please login as admin'}),401
         if not request.content_type == 'application/json': 
             return jsonify({'error':'Wrong content-type'}),400
         if not prodt_name or not prodt_quantity or not prodt_price or not prodt_desc:
             return jsonify({'error':'required field cannot be empty'}),400
-        if re.search('[\s]',prodt_name) != None:
+        if re.search(r'[\s]',prodt_name) != None:
             return jsonify({'error':'poductname cannot contain spaces'}),400
         if not isinstance(prodt_price,int) or not isinstance(prodt_quantity,int) or prodt_price < 1 or prodt_quantity < 1:
             return jsonify({'error':'price or quantity must be a number and must be greater than 1'}),400
         product_to_update = db.select_a_product(productid)
         if product_to_update == None:
-            return jsonify({'error':'Product not found'}),404
+            return jsonify({'message':'Product not found'}),404
         if product_to_update[0] == int(productid):
             db.update_product(productid,prodt_name,prodt_quantity,prodt_price,prodt_desc)
-            return jsonify({'message':'Product successfuly updated'}),201
+            product_updated = db.select_a_product(productid)
+            product = {
+                        'Product_id':product_updated[0],
+                        'Product_name':product_updated[1],
+                        'Product_quantity':product_updated[2],
+                        'Product_price':product_updated[3],
+                        'Product_description':product_updated[4]
+                      }
+            return jsonify({'Product':product,'message':'Product successfully updated'}),201
     except Exception:
         return jsonify({'error':'Required field/s missing'}),400
-
 
 @product.route('/products/<productid>', methods=['DELETE'])
 @protected_route
 def delete_product(current_user,productid):
+    if check_admin(current_user) is True:
+        return jsonify({'error':'Access Denied. Please login as admin'}),401
     selected_product = db.select_a_product(productid)
     if selected_product == None:
-        return jsonify({'error':'Product not found'}),404
+        return jsonify({'message':'Product not found'}),404
     db.delete_product(productid)
-    return jsonify({'message':'Product was deleted successfuly'})
+    return jsonify({'message':'Product was deleted successfuly'}),200
